feat(visualizer): highlight relation lines on table hovered
Highlight relations lines when tables is hovered
