+++ packages/json-table-schema-visualizer/src/types/tablesInfoProviderValue.ts
@@ -2,4 +2,6 @@ export type ColsIndexesMap = Record<string, number>;
 
 export interface TablesInfoProviderValue {
   colsIndexes: ColsIndexesMap;
+  hoveredTableName: string | null;
+  setHoveredTableName: (tableName: string | null) => void;
 }
